+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add component to display chat, viewers and apps during a live in a panel
 - Add `consumer_site` in JWT token
 - Allow CORS headers configuration
+- Extract SVG pages from PDF files for shared live media in lambda-convert
 
 ### Changed